wakeupkoala/requirements.txt
DateTime==4.3
discord==1.0.1
discord.py==1.4.1
python-dotenv==0.14.0
requests==2.24.0
schedule==0.6.0